+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ## 1.2.2
 - Added Z80 register names as proposals for completions.
-- Fixed "Go to definition" for sjasm macros.
+- Fixed "Go to definition" for sjasmplus macros.
 
 
 ## 1.2.1